+++ programs/drift/src/instructions/keeper.rs
@@ -2094,21 +2094,19 @@ pub struct UpdatePerpBidAskTwap<'info> {
 pub struct UpdateUserQuoteAssetInsuranceStake<'info> {
     pub state: Box<Account<'info, State>>,
     #[account(
+        mut,
         seeds = [b"spot_market", 0_u16.to_le_bytes().as_ref()],
         bump
     )]
     pub spot_market: AccountLoader<'info, SpotMarket>,
     #[account(
         mut,
-        has_one = authority,
+        constraint = is_stats_for_if_stake(&insurance_fund_stake, &user_stats)?
     )]
     pub insurance_fund_stake: AccountLoader<'info, InsuranceFundStake>,
-    #[account(
-        mut,
-        has_one = authority,
-    )]
+    #[account(mut)]
     pub user_stats: AccountLoader<'info, UserStats>,
-    pub authority: Signer<'info>,
+    pub signer: Signer<'info>,
     #[account(
         mut,
         seeds = [b"insurance_fund_vault".as_ref(), 0_u16.to_le_bytes().as_ref()],
@@ -2121,21 +2119,19 @@ pub struct UpdateUserQuoteAssetInsuranceStake<'info> {
 pub struct UpdateUserGovTokenInsuranceStake<'info> {
     pub state: Box<Account<'info, State>>,
     #[account(
+        mut,
         seeds = [b"spot_market", 15_u16.to_le_bytes().as_ref()],
         bump
     )]
     pub spot_market: AccountLoader<'info, SpotMarket>,
     #[account(
         mut,
-        has_one = authority,
+        constraint = is_stats_for_if_stake(&insurance_fund_stake, &user_stats)?
     )]
     pub insurance_fund_stake: AccountLoader<'info, InsuranceFundStake>,
-    #[account(
-        mut,
-        has_one = authority,
-    )]
+    #[account(mut)]
     pub user_stats: AccountLoader<'info, UserStats>,
-    pub authority: Signer<'info>,
+    pub signer: Signer<'info>,
     #[account(
         mut,
         seeds = [b"insurance_fund_vault".as_ref(), 15_u16.to_le_bytes().as_ref()],