+++ CHANGELOG.md
@@ -9,7 +9,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Features
 
+- program: add logging to pyth pull updates ([#1189](https://github.com/drift-labs/protocol-v2/pull/1189))
+
 ### Fixes
+
+### Breaking
+
+## [2.92.0] - 2024-08-12
+
+### Features
+
+- program: init prediction markets ([#1152](https://github.com/drift-labs/protocol-v2/pull/1152))
+
+### Fixes
+
+- program: make updateUserQuoteAssetInsuranceStake permissionless ([#1187](https://github.com/drift-labs/protocol-v2/pull/1187))
+- program: fix openbook v2 oom issue ([#1186](https://github.com/drift-labs/protocol-v2/pull/1186))
 
 ### Breaking